Made code more modular and made the drop down into a component
