+++ client/src/App.tsx
@@ -94,7 +94,7 @@ function App() {
                 Chat to Agent
               </Link>
 
-              {/* <Link
+              <Link
                 to="/token-launch"
                 className={`text-lg font-semibold transition duration-300 px-4 py-2 rounded-lg
                   ${
@@ -104,7 +104,7 @@ function App() {
                   }`}
               >
                 Token Launch
-              </Link> */}
+              </Link>
               {/*           <Link
             to="/social-dashboard"
             className={`text-lg font-semibold transition duration-300 px-4 py-2 rounded-lg
