--- server_node/packages/pumpfun/example/basic/.keys/mint.json
@@ -1,1 +1,0 @@
-{"secretKey":"57M1DkEtC4fuX3Q5hUBztvuaQAs23SP8x3RQcpGqHvTTHwwZ5DRvWaSGchjWgM9ynrGopo4SN4KRDreLo4HWzo3z","publicKey":"4rCXFNGs3o2PytWRGnHp6qJRwNE7eadnibMRPzqoP62Y"}
