+++ server_node/packages/pumpfun/example/basic/.keys/test-account.json
@@ -1,1 +1,1 @@
-{"secretKey":"2DVfiLqbUdQiZk7Cthncd4kHhjMsGF8usBnkxvpiYwHyAGQnXunDR2SCkHy1bnJSGn8Mbbgt7SxyYNgrsTnfDPSQ","publicKey":"DiN28eJtckQphwZsHGb1j3a9mhxGDddTy8WZfhZPibYc"}
+{"secretKey":"2UY4i4jtVFvyBGzcLcuuUN2jbvQYh5SrGgBTq1Z4yWu5o4EXJvZBLLJAi8V7nUUjXXEwX15QNzEoepDgn7G13bZ2","publicKey":"HFTKPREjYoVCbmGKJ6i2bjwXAvWR4Bsq1PTkyTduy5rG"}
